tests/wpt/web-platform-tests/html/cross-origin-opener-policy/access-reporting/opener-accessed_openee-coop.https.html

<title>
  COOP reports are sent when the openee used COOP+COEP and then tries to
  access its opener.
</title>
<meta name=timeout content=long>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/common/get-host-info.sub.js></script>
<script src="/common/utils.js"></script>
<script src="./resources/dispatcher.js"></script>
<script>

const directory = "/html/cross-origin-opener-policy/access-reporting";
const executor_path = directory + "/resources/executor.html?pipe=";
const cross_origin = get_host_info().HTTPS_REMOTE_ORIGIN;
const coep_header = '|header(Cross-Origin-Embedder-Policy,require-corp)';

let operation = [
//[property      , operation                   ] ,
  ["blur"        , 'opener.blur()'             ] ,
  ["close"       , 'opener.close()'            ] ,
  ["closed"      , 'opener.closed'             ] ,
  ["focus"       , 'opener.focus()'            ] ,
  ["frames"      , 'opener.frames'             ] ,
  ["indexed"     , 'opener[0]'                 ] ,
  ["indexed"     , 'opener[0] = ""'            ] ,
  ["length"      , 'opener.length'             ] ,
  ["location"    , 'opener.location'           ] ,
  ["location"    , 'opener.location = "#"'     ] ,
  ["named"       , 'opener["test"]'            ] ,
  ["named"       , 'opener["test"] = ""'       ] ,
  ["opener"      , 'opener.opener'             ] ,
  ["opener"      , 'opener.opener = ""'        ] ,
  ["postMessage" , 'opener.postMessage("")'    ] ,
  ["postMessage" , 'opener.postMessage("", "")'] ,
  ["self"        , 'opener.self'               ] ,
  ["top"         , 'opener.top'                ] ,
  ["window"      , 'opener.window'             ] ,
];

operation.forEach(([property, op]) => {
  promise_test(async t => {
    const report_token = token();
    const executor_token = token();
    const callback_token = token();

    const reportTo = reportToHeaders(report_token);
    const openee_url = cross_origin + executor_path +
      reportTo.header + reportTo.coopSameOriginHeader + coep_header +
      `&uuid=${executor_token}`;
    const openee = window.open(openee_url);
    t.add_cleanup(() => send(executor_token, "window.close()"));

    // 1. Skip the first report about the opener breakage.
    let report_1 = await receive(report_token);
    assert_not_equals(report_1, "timeout",
      "Receive navigation-to-document report");
    report_1 = JSON.parse(report_1);
    assert_equals(report_1.length, 1);
    assert_equals(report_1[0].type, "coop");
    assert_equals(report_1[0].body["violation-type"], "navigation-to-document");
    assert_equals(report_1[0].body["disposition"], "enforce");

    // 3. Try to access the opener. A report is sent, because of COOP+COEP.
    send(executor_token, `
      try {${op}} catch(e) {}
      send("${callback_token}", "Done");
    `);
    let reply = await receive(callback_token);
    assert_equals(reply, "Done");

    // 4. A COOP access reports must be sent as a result of (3).
    let report_2 = await receive(report_token);
    assert_not_equals(report_2, "timeout", "Received access report");
    report_2 = JSON.parse(report_2);

    assert_equals(report_2.length, 1);
    assert_equals(report_2[0].type, "coop");
    assert_equals(report_2[0].url, openee_url.replace(/"/g, '%22'));
    assert_equals(report_2[0].body["violation-type"], "access-from-coop-page");
    assert_equals(report_2[0].body["disposition"], "enforce");
    assert_equals(report_2[0].body["effective-policy"], "same-origin-plus-coep");
    assert_equals(report_2[0].body["property"], property);
    assert_equals(report_2[0].body["source-file"], undefined);
    assert_equals(report_2[0].body["lineno"], undefined);
    assert_equals(report_2[0].body["colno"], undefined);
  }, `${op}`);
});

</script>
